# Improve overall performance and use less memory