--- a/files/files.go
+++ b/files/files.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"strconv"
 	"strings"
 )
 
@@ -41,7 +40,7 @@ func NewFilename(name, extension string) (string, error) {
 	}
 	// Resolve duplicates
 	for i := 0; Exists(formatted); i++ {
-		formatted, err = format("("+strconv.Itoa(i)+") "+name, extension)
+		formatted, err = format(fmt.Sprintf("(%d) %s", i, name), extension)
 		if err != nil {
 			return "", fmt.Errorf("%w: failed to create filename (name=%s,ext=%s)", err, name, extension)
 		}
@@ -51,14 +50,13 @@ func NewFilename(name, extension string) (string, error) {
 
 // Exists returns whether the file exists.
 func Exists(filename string) bool {
-	f, err := os.Stat(filename)
-	if err != nil {
-		return os.IsExist(err)
+	if _, err := os.Stat(filename); err != nil {
+		return false
 	}
-	return !f.IsDir()
+	return true
 }
 
-const MaxFilenameLength = 200
+const MaxFilenameLength = 255
 
 // format ensures that the filename is valid for NTFS and has the right extension.
 func format(filename, extension string) (string, error) {
@@ -77,19 +75,17 @@ func format(filename, extension string) (string, error) {
 		requiredLength := MaxFilenameLength - len(extension) - 1
 		filename = filename[:requiredLength]
 	}
-	return filename + "." + extension, nil
+	return fmt.Sprintf("%s.%s", filename, extension), nil
 }
+
+var forbiddenChars = []string{"/", "<", ">", ":", "\"", "\\", "|", "?", "*"}
 
 // removeForbiddenChars removes invalid characters for Linux/Windows filenames.
 func removeForbiddenChars(name string) string {
-	var (
-		forbiddenChars = []string{"/", "<", ">", ":", "\"", "\\", "|", "?", "*", "(", ")"}
-		result         = name
-	)
 	for _, c := range forbiddenChars {
-		result = strings.ReplaceAll(result, c, "")
+		name = strings.ReplaceAll(name, c, "")
 	}
-	return result
+	return name
 }
 
 // NavigateTo moves to the provided directory and creates it if necessary.