This is the comparison to the current implementation (old):

```
name            old time/op    new time/op    delta
Download1-12       491ms ± 0%     252ms ± 0%   ~     (p=1.000 n=1+1)
Download10-12      2.05s ± 0%     1.40s ± 0%   ~     (p=1.000 n=1+1)
Download25-12      3.79s ± 0%     3.86s ± 0%   ~     (p=1.000 n=1+1)
Download50-12      5.31s ± 0%     4.80s ± 0%   ~     (p=1.000 n=1+1)
Download100-12     7.75s ± 0%     7.37s ± 0%   ~     (p=1.000 n=1+1)

name            old alloc/op   new alloc/op   delta
Download1-12      7.04MB ± 0%    6.75MB ± 0%   ~     (p=1.000 n=1+1)
Download10-12      110MB ± 0%     109MB ± 0%   ~     (p=1.000 n=1+1)
Download25-12      204MB ± 0%     202MB ± 0%   ~     (p=1.000 n=1+1)
Download50-12      553MB ± 0%     549MB ± 0%   ~     (p=1.000 n=1+1)
Download100-12    1.06GB ± 0%    1.05GB ± 0%   ~     (p=1.000 n=1+1)

name            old allocs/op  new allocs/op  delta
Download1-12       3.30k ± 0%     0.46k ± 0%   ~     (p=1.000 n=1+1)
Download10-12      20.3k ± 0%     14.1k ± 0%   ~     (p=1.000 n=1+1)
Download25-12      45.4k ± 0%     27.3k ± 0%   ~     (p=1.000 n=1+1)
Download50-12      89.1k ± 0%     54.1k ± 0%   ~     (p=1.000 n=1+1)
Download100-12      174k ± 0%       69k ± 0%   ~     (p=1.000 n=1+1)
```